# docs/openapi.yaml
openapi: "3.0.2"
info:
  title: API Title
  version: "1.0"
servers:
  - url: /api
paths:
...
  /articles/{slug}:
    get:
      tags:
        - Articles
      summary: 記事単一取得
      description: 単一記事を取得します
      operationId: GetArticle
      parameters:
        - name: slug
          in: path
          description: 記事の slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleArticleResponse'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
    put:
      tags:
        - Articles
      summary: 記事更新
      description: 記事を更新します
      operationId: UpdateArticle
      parameters:
        - name: slug
          in: path
          description: 記事の slug
          required: true
          schema:
            type: string
      requestBody:
        description: 更新用の記事
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArticleRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleArticleResponse'
        '400':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericErrorModel'
    delete:
      tags:
        - Articles
      summary: 記事削除
      description: 記事を削除します
      operationId: DeleteArticle
      parameters:
        - name: slug
          in: path
          description: 記事の slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content: { }
components:
  schemas:
    GenericErrorModel:
      required:
        - errors
      type: object
      properties:
        errors:
          required:
            - body
          type: object
          properties:
            body:
              type: array
              items:
                type: string
    Article:
      required:
        - slug
        - body
        - description
        - title
      type: object
      properties:
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        body:
          type: string
    SingleArticleResponse:
      required:
        - article
      type: object
      properties:
        article:
          $ref: '#/components/schemas/Article'
...
    UpdateArticle:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        body:
          type: string
    UpdateArticleRequest:
      required:
        - article
      type: object
      properties:
        article:
          $ref: '#/components/schemas/UpdateArticle'